% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootmat}
\alias{bootmat}
\title{Function to bootstrap matrix}
\usage{
bootmat(inmat, centroids, r, seed = NULL)
}
\arguments{
\item{inmat}{An input matrix with features as rows and samples as columns}

\item{centroids}{A character vector indicating the centroids}

\item{r}{A numeric correlation threshold}

\item{seed}{An integer to set the random seed}
}
\value{
A matrix describing which edges were significant in the bootstrapped
matrix according to the r correlation threshold provided
}
\description{
Function to bootstrap matrix
}
