% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fcor}
\alias{fcor}
\title{A fast correlation function}
\usage{
fcor(inmat, centroids, r)
}
\arguments{
\item{inmat}{An input matrix with features as rows and samples as columns}

\item{centroids}{A character vector indicating the centroids}

\item{r}{A numeric correlation threshold}
}
\value{
A matrix describing which edges were significant in the input matrix
matrix according to the r correlation threshold provided
}
\description{
A fast correlation function
}
