% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_regulon}
\alias{filter_regulon}
\title{filter_regulon - Filter a Regulon Object}
\usage{
filter_regulon(
  regulon,
  tf_list = NULL,
  likelihood_threshold = 0,
  correlation_threshold = 0
)
}
\arguments{
\item{regulon}{A regulon object, as generated by corto()}

\item{tf_list}{A character vector of centroid names to keep. Default is
NULL, i.e. all centroids}

\item{likelihood_threshold}{Minimum edge likelihood, between 0 and 1.
Default is 0}

\item{correlation_threshold}{Minimum absolute correlation. Default is 0}
}
\value{
A regulon object, or NULL if no edge survives
}
\description{
Subsets a regulon by centroid, by edge likelihood and by absolute
correlation. Centroids left without targets are removed.
}
\examples{
load(system.file("extdata","inmat.rda",package="corto",mustWork=TRUE))
load(system.file("extdata","centroids.rda",package="corto",mustWork=TRUE))
regulon<-corto(inmat,centroids=centroids,nbootstraps=10,p=1e-30,nthreads=2)
filtered<-filter_regulon(regulon,likelihood_threshold=0.5,correlation_threshold=0.9)
length(filtered)
}
