% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{funboot}
\alias{funboot}
\title{Function to process bootstraps into edges}
\usage{
funboot(seed = 0, inmat, centroids, r, selected_edges, targets)
}
\arguments{
\item{seed}{An integer to set the random seed}

\item{inmat}{An input matrix with features as rows and samples as columns}

\item{centroids}{A character vector indicating the centroids}

\item{r}{A numeric correlation threshold}

\item{selected_edges}{A character vector indicating which edges will be considered}

\item{targets}{A character vector indicating the targets}
}
\value{
A character vector with edges selected in this bootstrap
}
\description{
Function to process bootstraps into edges
}
