% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getregulon}
\alias{getregulon}
\title{getregulon - Flatten a Regulon Object into a Table}
\usage{
getregulon(
  regulon,
  tf_list = NULL,
  likelihood_threshold = 0,
  export = FALSE,
  filename = "regulon.tsv"
)
}
\arguments{
\item{regulon}{A regulon object, as generated by corto()}

\item{tf_list}{A character vector of centroid names to keep. Default is
NULL, i.e. all centroids}

\item{likelihood_threshold}{Minimum edge likelihood, between 0 and 1.
Default is 0}

\item{export}{A logical value. If TRUE the table is also written to file.
Default is FALSE}

\item{filename}{The file to write when export is TRUE. Default is
"regulon.tsv"}
}
\value{
A data frame with columns tf, target, correlation and likelihood
}
\description{
Converts a regulon into a data frame of edges, one row per
centroid-target pair, optionally written to a tab-separated file.
}
\examples{
load(system.file("extdata","inmat.rda",package="corto",mustWork=TRUE))
load(system.file("extdata","centroids.rda",package="corto",mustWork=TRUE))
regulon<-corto(inmat,centroids=centroids,nbootstraps=10,p=1e-30,nthreads=2)
edges<-getregulon(regulon)
head(edges)
}
