% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{p2r}
\alias{p2r}
\title{p2r Convert a P-value to the corresponding Correlation Coefficient}
\usage{
p2r(p, n)
}
\arguments{
\item{p}{the p-value}

\item{n}{the number of samples}
}
\value{
a correlation coefficient
}
\description{
p2r Convert a P-value to the corresponding Correlation Coefficient
}
\examples{
p2r(p=0.08,n=20)
}
