% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{p2z}
\alias{p2z}
\title{p2z}
\usage{
p2z(p)
}
\arguments{
\item{p}{a p-value}
}
\value{
z a Z score
}
\description{
This function gives a gaussian Z-score corresponding to the provided p-value
Careful: sign is not provided
}
\examples{
p<-0.05
p2z(p)
}
