% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{r2p}
\alias{r2p}
\title{r2p Convert Correlation Coefficient to P-value}
\usage{
r2p(r, n)
}
\arguments{
\item{r}{the correlation coefficient}

\item{n}{the number of samples}
}
\value{
a numeric p-value
}
\description{
r2p Convert Correlation Coefficient to P-value
}
\examples{
r2p(r=0.4,n=20) # 0.08
}
