% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{slice}
\alias{slice}
\title{Slice}
\usage{
slice(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
prints it
}
\description{
This function prints a slice of a matrix
}
\examples{
set.seed(1)
example<-matrix(rnorm(1000),nrow=100,ncol=10)
slice(example)
}
