% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{stouffer}
\alias{stouffer}
\title{Stouffer integration of Z scores}
\usage{
stouffer(x)
}
\arguments{
\item{x}{a vector of Z scores}
}
\value{
Z an integrated Z score
}
\description{
This function gives a gaussian Z-score corresponding to the provided p-value
Careful: sign is not provided
}
\examples{
zs<-c(1,3,5,2,3)
stouffer(zs)
}
