% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{z2p}
\alias{z2p}
\title{z2p}
\usage{
z2p(z)
}
\arguments{
\item{z}{a Z score}
}
\value{
a p-value
}
\description{
This function gives a gaussian p-value corresponding to the provided Z-score
}
\examples{
z<-1.96
z2p(z)
}
