% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulon_utils.R
\name{filter_regulon}
\alias{filter_regulon}
\title{Filter a regulon object by TF list, likelihood and correlation thresholds}
\usage{
filter_regulon(
  regulon,
  tf_list = names(regulon),
  likelihood_threshold = 0,
  correlation_threshold = 0
)
}
\arguments{
\item{regulon}{A regulon object returned by corto()}

\item{tf_list}{Character vector with names of transcription factors to retain}

\item{likelihood_threshold}{Minimum likelihood (0 to 1) to keep an interaction}

\item{correlation_threshold}{Minimum absolute correlation to keep an interaction}
}
\value{
A filtered regulon object or NULL if empty
}
\description{
Filter a regulon object by TF list, likelihood and correlation thresholds
}
