% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mra.R
\name{mraplot}
\alias{mraplot}
\title{Plot a master regulator analysis}
\usage{
mraplot(mraobj, mrs = NULL)
}
\arguments{
\item{mraobj}{The input object, output of the function mra}

\item{mrs}{Either a numeric value indicating how many MRs to show, sorted by
significance, or a character vector specifying which TFs to show}
}
\value{
A plot is generated
}
\description{
Plotting function for master regulator analysis performed by the _mra_ function
}
